pyproject.toml
[tool.poetry]
name = "clonegpt"
version = "0.1.0"
description = ""
authors = ["MOREL Mathias <mathias_morel@ens.univ-artois.fr>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "2.3.3"
openai = "^1.50.2"
python-dotenv = "1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
